Speed up audio processing
As audio is received, it's placed into a large ring buffer.

Every 100ms, some audio, the last 2, 4, and 8, and 16 seconds, are
copied out into four processing buffers.

This means each 100ms that goes into the buffer is copied out 300 times
before it's no longer used.  So it's more efficient to do processing on
the audio as it goes in, since it only need be done once instead of 300
times.

Some of the processing that was done when copying the audio out of ring
buffer was to add both channels together and to drop every other sample
in light mode.  Do this instead when the audio is added to the buffer.

There are some other improvements, besides doing this 300 times less
often.

Instead of de-interleaving the audio into two buffers, one for each
channel, we can write the combined audio into one buffer.  This touches
half as many cache lines, since the two buffers were far apart, they
would never in in the same cache line.

In one channel mode, we can just copy the data with memcpy(), which is
much faster than a loop.  This avoids needing to duplicate the one
channel twice to create two buffers only to add the two buffers back
together.

Since the data in the ring buffer is already summed and sample dropped,
we can copy it out with memcpy(), and this operation, which again
happens 300 times more often, is now much faster still.

When switching between light and normal mode, the audio in the ring
buffer, which would be decimated or not already based on mode, is
cleared so that it isn't processed incorrectly.

After doing this, the time spent getting the audio into the algorithm
basically drops off the code execution profile.