Faster version of tmean()
A profile shows a significant portion of the total runtime is spent in
tmean().  This function finds the mean of the lower four quintiles of a
list.  I.e, the mean of the values in the 0th to 80th percentile and
excluding the values in the 80th to 100th percentile.  The exact break
point is about 80, but depends on the number of items in the list, as a
single item can't be split.  The lists typically have 5 to 80 elements,
depending on bph and the processing buffer size.

It did this by sorting the list when it had 16 or fewer elements and
with larger lists it used an approximation.  This was done by sorting
the first 16 elements and finding the 4th quintile of those, to serve as
an estimate of the entire list's fourth quintile.  Then the actual rank
of that estimate was found by examining the full list.  If it was
between the 70th and 90th percentile, the mean of elements less than the
estimate was used.  If not, the entire list was sorted to find the exact
elements in to be used, in the same way as lists of 16 elements or
fewer.

The complexity of this sorting was O(n log n).  The estimate was used
about 80% of the time when the list was long enough to use that
algorithm.

The algorithm used here is a "quickselect" to partition the list into
quintiles 1-4 and quintile 5.  This is like a quicksort, but average
case complexity is just O(n), because it doesn't need to sort the entire
list, and instead need only partition the one branch that the 4th
quintile break point is in.  This ends up being faster than the previous
approximation algorithm even when the estimate was accepted, and even
more so when the estimate was rejected and the entire list needed to be
sorted.  So there is no gain to use the approximation for large lists.

For small lists, < 10 elements, simpler algorithms are used.  The
algorithms are to sum the last whilst simultaneously finding the least
(or two least) elements, which are then subtracted from the sum.  These
algorithms are technically O(n^2), but are faster for this small size by
virtue of the simplicity and sequentialness of their implementations.
With more than 10 elements it would be necessary to find the smallest
three elements and this did not benchmark as faster.