Use quickselect to find median
There were three places that found the median by sorting a list and
taking the middle element.  The quickselect function is faster.

None of this code appears to depend on the list being sorted after
finding the median.  Quickselect doesn't do this, which is what allows
it to be O(n) vs O(n log n).

It might be faster still to use a median of medians algorithm for the
cases where the exact median is not necessary and an approximate median
will do.