Run high pass filter in audio thread
This moves the high pass filter out of prepare_data() and into the audio
callback.

prepare_data() will process the same data many times, up to 300 times,
as it is repeatedly called to processes buffers that mostly overlap the
last set of buffers.

Running the HPF on incoming audio will process about 300x less data
without changing the result.  This produces about a 4% overall speed
improvement.

In the next step in the audio processing, the noise filter, each sample
depends on every other sample in the buffer, so all samples need to be
re-processed each time the buffers change.  So it's not possible to do
more processing past the HPF as the audio is received.

I also benchmarked code that would apply the HPF while coping the data
out of from the portaudio buffer to the ring buffer.  I.e., a biquad
filter function that did not modify the data in place, but instead had
separate input and output pointers.  Dealing with the audio
combinations, mono/stereo, normal/light, took more code and there was no
significant speed improvement.

I also tried moving the HPF out of the audio callback and into the
compute thread, while still applying the filter only once to new audio,
but this was slower, doesn't spread the load across multple CPUS as
well, and was more complex.