Fix race with light mode switch and audio callback
When light mode is toggled, it's possible for the buffer pointer or
timestamp to get messed up if it happens while the callback is running.

The callback will read the current values when it starts, then light
mode switch resets them to zero, and then when the callback finishes it
updates them based on the original values, effectvely undoing to reset
to zero.

This doesn't cause a crash, but it does mean the timestamp isn't quite
right.  It also will cause problems if any code wants to process the
audio data as it arrives.

It could cause a single callback to be processed with the incorrect high
pass filter.

The easiest way to fix this is to pause the input stream while doing the
mode switch.  It could be fixed by making the entire callback a critical
section or having the mode switch done asyncrounsly via the callback,
but this makes the callback more complex and adds overhead.  It seems
better to add the extra work to the mode switch rather than the
callback, since the former is a rare operation that isn't performance
critical while the latter is.