Fix issue preventing BPH below 14400 from working
Periods (two beats) longer than 500 ms were rejected as too long.  This
effectively limits the low end of BPH to 14400.  Any slower and the
correct period length is greater than 500 ms, e.g. 12000 BPH is a 600 ms
period.

Change the limit to be 20% slower than the nominal value for the BPH.
If the BPH is set, then use that, when BPH is guessed, use MIN_BPH.

A MIN_BPH of 12000 will increases the max period to 720 ms when
guessing.  It could be longer if one were able to set a slower than
MIN_BPH period length.