Break up the output panel GUI creation function
It's quite large.  Split it into new -functions that create the various
parts of the output panel.

The waveforms (tic, toc, period, debug), the paper strip chart and its
controls, and the window under the output display that contains them,
each get their own function.

There is some minimal support for a vertical vs horizontal paperstrip
chart, but it's not used yet.