Use resizble paned widget for paperstrip and waveforms
This lets the size of the paperstrip vs the waveforms be adjusted by
sliding the pane between them.

To get the layout right, we need to give the waveforms a minimum size.
300x150 seems about the minimum.  It should look basically the same as
it did before.

Each snapshot has its own unique pane position.  Adjusting one tab in
the notebook doesn't adjust the other tabs.  It might be nice if that
weren't the case and they were all synced.