Add vertical/horizontal paperstrip option
This allow changing the layout so the paperstrip is horizontal,
scrolling to the left, with the three waveforms arraigned horizontally
below the paperstrip.

The effect is to make the paperstrip longer and give it more space.

For many tasks the paper strip is much more useful than the waveform
displays.  This allows adjusting the layout to be more useful in those
cases.

A new checkbox in the menu is added to control this.  It can be changed
on the fly and the widgets will be re-arraigned.

A config file option, "vertical_paperstrip", is added to remember the
current setting.  Default vertical layout to true if not present in
config file.

If multiple snapshots are present, the change of orientation will affect
all of them, rather that just the current viewed one.  It would be
possible to change this.