Move snapshot display parameters to their own struct
Create a new struct for the output_panel snapshot display parameters.
Instead of being mixed into the snapshot data from the computer, there
will just be a pointer to the display parameter struct in the snapshot.

The clone and delete snapshot functions will clone and delete the
display parameters, if they exist.  The computer's snapshots don't have
display parameters so the pointer will just be NULL for them.

Each time a new snapshot is passed from the computer to the interface,
the pointer to the display parameters can be kept so that each parameter
doesn't need to be copied.

paperstrip_draw_event() will allocate a new display parameter when it
gets a snapshot for the first time.

This will make it a lot easier to add new display parameters just by
adding them to the struct.  And they can be initialized in the output
panel code that uses them and knows what they should be.  The computer
doesn't need to know anything about them.

This does break the savefile format slightly since the struct field
names are coded into the savefile.  Maybe that design has some drawbacks
when it comes to backward compatibility.  In this case the trace
centering parameter is the only one broken and it doesn't really matter.