Size fonts based on drawing area
Since the space allocation between the paperstrip and waveforms can be
changed now, using just the main window size to control font size
doesn't work as well.  E.g., a small waveform in a large app window that
is mostly paperstrip will still get a large font, because the app window
is large.

Also adjust font spacing some.  Use the vertical size of the font to
adjust legend in paperstrip and provide consistent margins in waveforms.