Add zoom slider for paperstrip
Add a pair of buttons that overlay the paperstrip.  One is a Gtk slider
button which allows zooming in and out with a slider.  The other will
reset to the default zoom.

This zoom is just zoom the chart in the horizontal direction, i.e. error
in each beat.  Not vertically, which would show more or fewer beats at
once.

The zoom ranges from the 1.0 to 0.01.  E.g., for a 10 beat/sec movement,
the chart's width will be at most 100 ms, an entire beat, to 10 ms, the
default, down to 1 ms.

The forced switch from 0.1 in normal mode to 0.01 in calibration mode is
removed.  If one wants to zoom in for calibration, then one is free to
do that.

Due to the way the paperstrip is drawn, only zoom values that are exact
integer fractions will work.  E.g., 0.5 = 1/2, 0.2 = 1/5, and 0.10 =
1/10, will all work.  But 0.12 = 1/8.333 will not, because 8⅓ is not an
integer.  This makes the initial zoom jumps quite large.