Redo paperstrip display, new dot algorithm
Re-write most of the code for generating the paperstrip.

A new algorithm is used to place the dots which has an advantage in that
it can display correctly with any level of zoom, not just at integer
scales.

What we do is find each event's relative position from the one before
it.  A relative offset of 0 would mean the event is exactly a multiple
of the beat time from the previous event.  A positive or negative value
means it is earlier or later than expected.  We can then scale this
offset by any value.

Since only the relative position between events is used there is no
longer a problem when the accumulated offset wraps around a beat length
but not the chart width.

With all event positions relative, we need some absolute position from
which to start.  This is done by saving the absolute offset of the most
recent event as an anchor.  The next display will place the previous
anchor at the same absolute offset and update the anchor to the new most
recent event.  This way the dots do not shift side-to-side as they
scroll.

Having done this, much of the other code relating to the paperstrip
becomes simpler and there is a net reduction in lines of code.

The dots are placed in nearly the same positions as before, with one
significant difference.  The slope is reversed, i.e. a movement running
fast will slope up and to the left.  This works better for the math, and
appears to match what other timegraphers do.