Dynamically allocate sound buffer
The buffer used for audio data is a static size based on what's needed
for 32 seconds of data at the fixed 44.1kHz sampling rate.  This doesn't
work as well if one wants to change the sample rate and possibly support
high rates.

Dynamically allocate it.  If the sampling rate were to increase,
re-allocate a larger buffer for it.