Save audio device and rate selection to config
This will save the audio device and rate selected to the config file.

Initially the saved device and rate are both codes for "default", so the
current behavior of using the Portaudio default device at 44.1 kHz
remains.  Only if the audio setup dialog is used to select something
different, and that selection is started successfully, does a specific
device and rate get saved.

Additionally, selecting the default device in audio setup is still saved
as "default", rather than the specific device that happens to be the
current default.

If the saved device is not available when starting up it will
automatically revert to the default.