Hide zoom reset button when at original scale
When at the default zoom scale, hide the zoom reset button.  No need to
have it on the screen since it won't do anything.  And it tells you that
the scale is already at default.

I noticed the zoom reset button in Chrome disappeared when at 100% and
liked this feature.