Processes only need steps for large speedup
I measured this as reducing CPU usage from ~78% to ~45%.  Almost
doubling the speed, which is expected as described below.

Tg processes the audio in a series of steps that double in size:  2, 4,
8, and then 16 seconds long.  This is the one to four dots shown in the
display.

Tg starts at step 1 (2 seconds) and goes up, stopping if a step doesn't
pass.  Data from smaller steps isn't used if a larger step passes.

This means when running at a constant step 4 with good signal, CPU usage
is almost double what it needs to be, since steps 1-3 are processed and
unused and add up to 14 seconds, almost as much as step 4's 16 seconds.

Change this to start at the previous iteration's step.  With good
signal, only step 4 will be processed and CPU usage is cut almost in
half.  If the previous step fails, it will try smaller steps.  If it
passes and is not yet the top step, it will try larger steps.

End result should end up on the same step as before, but get there
sooner.  It could be slower if the step drops a lot, e.g. from 4 to 0,
but this happens far less often than the step saying the same or nearly
the same.

To do this, I moved the step logic out of analyze_pa_data() and entirely
into compute_update().  analyze_pa_data() will now just processes one
step and compute_update() decides which step(s).  Previously,
analyze_pa_data() did all steps and compute_update() decided which step
to actually use.

There is a small change to the algorithm.  To be good, a step needs to
pass a number of changes done in process().  Then there is one more
check, of sigma vs period, done in compute_update().

Previously a step didn't need to pass the sigma check and it still
counted enough to increase last_tic and show up as a signal dot in the
display.  But the data wasn't actually used unless it passed to sigma
check too.

I no longer keep track of "partially" passing steps like this.  Either
it passes all checks, including the sigma check, or not.  last_tic and
signal level only count fully passing steps.

I see no practical difference in my tests, but I think it could show up
with some kind of marginal signal that has a high error in period
estimation with the longer steps.