Add an RPM spec file for building on Fedora, etc.
This works on Fedora, Redhat, CentOS, etc.  It also works for
autobuilding on COPR.

It will regenerate the autoconf code, which allows it to work from the
git tree, i.e. without the configure script that is only in the release
tarballs.

The version number will be injected automatically, both when building
with autoconf and on COPR, which was quite a challenge.

When the configure script is run, it will preprocess the spec file and
an include file for it and inject the version.  The spec file's macro
logic will detect that this has been done and use it.

But on COPR the SRPM is created by rpkg without running configure, and
there is no way to add extra commands to do it, so this won't work.
However, rpkg has its own preprocessing systems for spec files and so
that is used to inject the version number.  The spec file macro logic
will use this if the autoconf injected version is not present.

It's also possible to specify the version manually on the rpmbuild
command line by defining the macro "version".

If pkgrel is defined then a release is being generated and the RPM
will be the release specified in pkgrel.

If pkgrel is not defined, then it's assumed a snapshot is being built
and the git date and commit are added following the Fedora guidelines
for snapshots.  Again, this needs to be done in two different ways to
support building from the git checkout directly or using COPR, which
builds binary RPMs from an SRPM rather than from a git checkout.