Allow configuring HPF cutoff frequency from audio setup dialog
Add a control to the setup dialog for the HPF cutoff frequency.  Also
save it to the config file.

Slider has ticks for Off and also for the default 3kHz frequency.  The
fill level of the slider is set to limit the cutoff to the Nyquist
frequency of the selected sampling rate.

To do that, I needed to get a callback on rate combobox change, and the
manual entry of the combobox made that a challenge.