Lower min BPH to 8100
Allow lower BPH values.  Create another setting that affects the BPH
guessing code so it doesn't get confused by fractions of faster beats.

Lowest value is 2.25 Hz or 8100 BPH.  Lower rates do not work well or at
all.

Tg uses a minimum 2 second detection window and needs to see two periods
in this window.  So that it can see the 1st period matches the 2nd
period and the time between them is about what it should be from the
BPH.  In order to get two complete periods (two beats each) in a 2
second window, one needs 7200 BPH.  Below this and there might only be
one period in the window and nothing to detect.

But even at 7200, the beats at the beginning and end of the window might
be cut off, so it's necessary to go a little faster to be reliable.