Change one sample sigma estimation to 0
When there is just one period in the processing buffer it is not
possible to calculate the sample standard deviation (sigma).  In this
case, the period value was being used as the sigma estimate, which
effectively gives a huge sigma when there is 1 period in the buffer.

This results in the processing buffer being rejected as bad and a larger
buffer, which would have multiple periods, is never tried.

One period per buffer would happen with a combination of long period and
short buffer, e.g. a small BPH in light mode, since light mode uses
buffers half as long as normal mode.

Use 0 as the sigma value for 1 sample.  This means the buffer will pass
the sigma check and a larger buffer will be tried.